fix(graph-db-server tests): harden system-lifecycle fuzz against cross-sequence I3 false-positives
The 'tier-2-fuzz (fuzz-system-lifecycle)' check intermittently failed with
I3 ("node X has edge to Y which never existed") on early steps of late
sequences, even when the PR under test did not touch graph-db-server. Root
cause: between-sequence cleanup was best-effort — it DELETEd only nodes
still keyed in `tracked.nodesViaApi` at scan time, then slept 200ms.
Two leaks went unrecorded:

  1. Watcher-promoted file nodes (createFile -> chokidar -> graph) were
     never in `nodesViaApi`, so cleanup skipped them.
  2. The daemon's DELETE /graph/node/:id workflow returns 500 when its
     internal fs.unlink races with a watcher-driven unlink (ENOENT); the
     in-memory node persists.

Either way, leaked nodes survived into the next sequence. When that
sequence reset `tracked.deletedNodeIds`, any leaked node with an outgoing
edge to a previously-cleaned target tripped I3 because neither side of
its `targetExists || targetWasDeleted` predicate held.

Fix: make `deletedNodeIds` accumulate across sequences. `cleanupSequence`
records every entity it attempts to clean (API DELETE or unlink) before
the call, so I3 keeps treating leaked references as "deleted" even when
the daemon's DELETE failed silently. `resetForNextSequence` clears
per-sequence state but preserves the deletion ledger.

Also: split the 513-line test file into a thin driver plus six focused
helpers (prng / poll / daemon-http / types / actions / invariants / drain)
per the project's functional-design guidance and the 500-line gate.

Verified: 5/5 consecutive runs of the full 100-sequence test pass, ~72s each.

The daemon's non-idempotent DELETE workflow (500 on already-unlinked files)
is a separate pre-existing bug worth a follow-up ticket — but it does not
affect correctness of this test now that I3 no longer false-positives.
