+++ packages/systems/graph-db-server/tests/system-lifecycle-fuzz/prng.ts
@@ -1,0 +1,19 @@
+// Deterministic Mulberry32 PRNG + small picker utilities for fuzz replay.
+
+export function mulberry32(seed: number): () => number {
+  let a = seed
+  return function () {
+    a |= 0; a = (a + 0x6D2B79F5) | 0
+    let t = Math.imul(a ^ (a >>> 15), 1 | a)
+    t = (t + Math.imul(t ^ (t >>> 7), 61 | t)) ^ t
+    return ((t ^ (t >>> 14)) >>> 0) / 4294967296
+  }
+}
+
+export function pick<T>(rng: () => number, arr: readonly T[]): T {
+  return arr[Math.floor(rng() * arr.length)] as T
+}
+
+export function randInt(rng: () => number, min: number, max: number): number {
+  return min + Math.floor(rng() * (max - min + 1))
+}
