+++ packages/systems/graph-db-server/tests/system-lifecycle.fuzz.test.ts
@@ -1,410 +1,16 @@
-import { mkdir, mkdtemp, rm, writeFile, unlink, stat, readFile } from 'node:fs/promises'
+import { mkdir, mkdtemp, rm } from 'node:fs/promises'
 import { tmpdir } from 'node:os'
 import path from 'node:path'
-import { afterEach, beforeEach, describe, expect, it } from 'vitest'
+import { afterEach, beforeEach, describe, it } from 'vitest'
 
 import { startDaemon, type DaemonHandle } from '../src/daemon/index.ts'
-
-// ---- Mulberry32 seeded PRNG (deterministic replay) ----
-
-function mulberry32(seed: number): () => number {
-  let a = seed
-  return function () {
-    a |= 0; a = (a + 0x6D2B79F5) | 0
-    let t = Math.imul(a ^ (a >>> 15), 1 | a)
-    t = (t + Math.imul(t ^ (t >>> 7), 61 | t)) ^ t
-    return ((t ^ (t >>> 14)) >>> 0) / 4294967296
-  }
-}
-
-function pick<T>(rng: () => number, arr: readonly T[]): T {
-  return arr[Math.floor(rng() * arr.length)] as T
-}
-
-function randInt(rng: () => number, min: number, max: number): number {
-  return min + Math.floor(rng() * (max - min + 1))
-}
-
-// ---- waitFor polling ----
-
-async function waitFor(read: () => Promise<boolean>, timeoutMs = 8000): Promise<void> {
-  const deadline = Date.now() + timeoutMs
-  while (Date.now() < deadline) {
-    if (await read()) return
-    await new Promise((resolve) => setTimeout(resolve, 50))
-  }
-  throw new Error('waitFor: condition not met before timeout')
-}
-
-// ---- Types ----
-
-type ActionType =
-  | 'createFile'
-  | 'deleteFile'
-  | 'upsertNodeDelta'
-  | 'upsertNodeWithEdges'
-  | 'updateExistingNode'
-  | 'deleteNodeDelta'
-  | 'deleteNodeEndpoint'
-  | 'getGraph'
-
-interface TrackedNode {
-  id: string
-  content: string
-  edges: string[]
-}
-
-interface TrackedState {
-  filesOnDisk: Map<string, string> // path -> expected content
-  nodesViaApi: Map<string, TrackedNode> // id -> tracked node
-  deletedNodeIds: Set<string> // nodes we explicitly deleted (daemon doesn't cascade incoming edges)
-}
-
-// ---- Helpers ----
-
-async function fetchJson(url: string, init?: RequestInit): Promise<{ status: number; body: unknown }> {
-  const res = await fetch(url, init)
-  const body = await res.json()
-  return { status: res.status, body }
-}
-
-interface GraphNode {
-  absoluteFilePathIsID: string
-  outgoingEdges: Array<{ targetId: string }>
-  contentWithoutYamlOrLinks: string
-}
-
-async function getGraph(baseUrl: string): Promise<{ nodes: Record<string, GraphNode> }> {
-  const { body } = await fetchJson(`${baseUrl}/graph`)
-  return body as { nodes: Record<string, GraphNode> }
-}
-
-async function fileExists(filePath: string): Promise<boolean> {
-  try {
-    await stat(filePath)
-    return true
-  } catch {
-    return false
-  }
-}
-
-// ---- Action generators ----
-
-function generateAction(
-  rng: () => number,
-  vault: string,
-  baseUrl: string,
-  tracked: TrackedState,
-  seqId: number,
-  stepId: number,
-): { type: ActionType; execute: () => Promise<void> } {
-  const candidates: ActionType[] = ['createFile', 'upsertNodeDelta', 'upsertNodeWithEdges', 'getGraph']
-
-  if (tracked.filesOnDisk.size > 0) candidates.push('deleteFile')
-  if (tracked.nodesViaApi.size > 0) {
-    candidates.push('deleteNodeDelta')
-    candidates.push('deleteNodeEndpoint')
-    candidates.push('updateExistingNode')
-  }
-
-  const actionType = pick(rng, candidates)
-
-  switch (actionType) {
-    case 'createFile': {
-      const fileName = `fuzz-${seqId}-${stepId}.md`
-      const filePath = path.join(vault, fileName)
-      const content = `# Fuzz ${seqId}-${stepId}\n\nContent seed=${seqId * 1000 + stepId}.\n`
-      const fileContent = `---\n---\n${content}`
-      return {
-        type: 'createFile',
-        execute: async () => {
-          await writeFile(filePath, fileContent, 'utf8')
-          tracked.filesOnDisk.set(filePath, content)
-        },
-      }
-    }
-
-    case 'deleteFile': {
-      const filePath = pick(rng, [...tracked.filesOnDisk.keys()])
-      return {
-        type: 'deleteFile',
-        execute: async () => {
-          if (await fileExists(filePath)) {
-            await unlink(filePath)
-          }
-          tracked.filesOnDisk.delete(filePath)
-        },
-      }
-    }
-
-    case 'upsertNodeDelta': {
-      const nodePath = path.join(vault, `fuzz-api-${seqId}-${stepId}.md`)
-      const content = `# API Node ${seqId}-${stepId}\n`
-      return {
-        type: 'upsertNodeDelta',
-        execute: async () => {
-          const delta = [
-            {
-              type: 'UpsertNode',
-              nodeToUpsert: {
-                kind: 'leaf',
-                absoluteFilePathIsID: nodePath,
-                outgoingEdges: [],
-                contentWithoutYamlOrLinks: content,
-                nodeUIMetadata: {
-                  color: { _tag: 'None' },
-                  position: { _tag: 'None' },
-                  additionalYAMLProps: {},
-                },
-              },
-              previousNode: { _tag: 'None' },
-            },
-          ]
-          const { status } = await fetchJson(`${baseUrl}/graph/delta`, {
-            method: 'POST',
-            headers: { 'content-type': 'application/json' },
-            body: JSON.stringify(delta),
-          })
-          expect(status).toBe(200)
-          tracked.nodesViaApi.set(nodePath, { id: nodePath, content, edges: [] })
-        },
-      }
-    }
-
-    case 'upsertNodeWithEdges': {
-      const nodePath = path.join(vault, `fuzz-edge-${seqId}-${stepId}.md`)
-      const content = `# Edge Node ${seqId}-${stepId}\n`
-      // Pick 1-3 existing nodes as edge targets
-      const existingIds = [...tracked.nodesViaApi.keys()]
-      const edgeCount = existingIds.length > 0 ? randInt(rng, 1, Math.min(3, existingIds.length)) : 0
-      const targets: string[] = []
-      for (let i = 0; i < edgeCount; i++) {
-        targets.push(pick(rng, existingIds))
-      }
-      const uniqueTargets = [...new Set(targets)]
-
-      return {
-        type: 'upsertNodeWithEdges',
-        execute: async () => {
-          const delta = [
-            {
-              type: 'UpsertNode',
-              nodeToUpsert: {
-                kind: 'leaf',
-                absoluteFilePathIsID: nodePath,
-                outgoingEdges: uniqueTargets.map((t) => ({ targetId: t, edgeLabel: '' })),
-                contentWithoutYamlOrLinks: content,
-                nodeUIMetadata: {
-                  color: { _tag: 'None' },
-                  position: { _tag: 'None' },
-                  additionalYAMLProps: {},
-                },
-              },
-              previousNode: { _tag: 'None' },
-            },
-          ]
-          const { status } = await fetchJson(`${baseUrl}/graph/delta`, {
-            method: 'POST',
-            headers: { 'content-type': 'application/json' },
-            body: JSON.stringify(delta),
-          })
-          expect(status).toBe(200)
-          tracked.nodesViaApi.set(nodePath, { id: nodePath, content, edges: uniqueTargets })
-        },
-      }
-    }
-
-    case 'updateExistingNode': {
-      const nodeId = pick(rng, [...tracked.nodesViaApi.keys()])
-      const existing = tracked.nodesViaApi.get(nodeId)!
-      const newContent = `# Updated ${seqId}-${stepId}\n`
-      // Optionally add/remove edges
-      const existingIds = [...tracked.nodesViaApi.keys()].filter((id) => id !== nodeId)
-      const newEdgeCount = existingIds.length > 0 ? randInt(rng, 0, Math.min(2, existingIds.length)) : 0
-      const newEdges: string[] = []
-      for (let i = 0; i < newEdgeCount; i++) {
-        newEdges.push(pick(rng, existingIds))
-      }
-      const uniqueEdges = [...new Set(newEdges)]
-
-      return {
-        type: 'updateExistingNode',
-        execute: async () => {
-          const graph = await getGraph(baseUrl)
-          if (!graph.nodes[nodeId]) {
-            tracked.nodesViaApi.delete(nodeId)
-            return
-          }
-          const delta = [
-            {
-              type: 'UpsertNode',
-              nodeToUpsert: {
-                kind: 'leaf',
-                absoluteFilePathIsID: nodeId,
-                outgoingEdges: uniqueEdges.map((t) => ({ targetId: t, edgeLabel: '' })),
-                contentWithoutYamlOrLinks: newContent,
-                nodeUIMetadata: {
-                  color: { _tag: 'None' },
-                  position: { _tag: 'None' },
-                  additionalYAMLProps: {},
-                },
-              },
-              previousNode: { _tag: 'Some', value: graph.nodes[nodeId] },
-            },
-          ]
-          const { status } = await fetchJson(`${baseUrl}/graph/delta`, {
-            method: 'POST',
-            headers: { 'content-type': 'application/json' },
-            body: JSON.stringify(delta),
-          })
-          expect(status).toBe(200)
-          tracked.nodesViaApi.set(nodeId, { id: nodeId, content: newContent, edges: uniqueEdges })
-        },
-      }
-    }
-
-    case 'deleteNodeDelta': {
-      const nodeId = pick(rng, [...tracked.nodesViaApi.keys()])
-      return {
-        type: 'deleteNodeDelta',
-        execute: async () => {
-          const graph = await getGraph(baseUrl)
-          if (!graph.nodes[nodeId]) {
-            tracked.nodesViaApi.delete(nodeId)
-            return
-          }
-          const delta = [
-            {
-              type: 'DeleteNode',
-              nodeId,
-              deletedNode: { _tag: 'Some', value: graph.nodes[nodeId] },
-            },
-          ]
-          const { status } = await fetchJson(`${baseUrl}/graph/delta`, {
-            method: 'POST',
-            headers: { 'content-type': 'application/json' },
-            body: JSON.stringify(delta),
-          })
-          expect(status).toBe(200)
-          tracked.nodesViaApi.delete(nodeId)
-          tracked.deletedNodeIds.add(nodeId)
-        },
-      }
-    }
-
-    case 'deleteNodeEndpoint': {
-      const nodeId = pick(rng, [...tracked.nodesViaApi.keys()])
-      return {
-        type: 'deleteNodeEndpoint',
-        execute: async () => {
-          const graph = await getGraph(baseUrl)
-          if (!graph.nodes[nodeId]) {
-            tracked.nodesViaApi.delete(nodeId)
-            return
-          }
-          const { status } = await fetchJson(
-            `${baseUrl}/graph/node/${encodeURIComponent(nodeId)}`,
-            { method: 'DELETE' },
-          )
-          expect(status).toBe(200)
-          tracked.nodesViaApi.delete(nodeId)
-          tracked.deletedNodeIds.add(nodeId)
-        },
-      }
-    }
-
-    case 'getGraph': {
-      return {
-        type: 'getGraph',
-        execute: async () => {
-          const graph = await getGraph(baseUrl)
-          expect(graph).toBeDefined()
-          expect(graph.nodes).toBeDefined()
-          expect(typeof graph.nodes).toBe('object')
-        },
-      }
-    }
-  }
-}
-
-// ---- Invariant checks ----
-
-async function assertInvariants(
-  baseUrl: string,
-  vault: string,
-  tracked: TrackedState,
-  ctx: string,
-): Promise<void> {
-  const graph = await getGraph(baseUrl)
-
-  // I1: GET /graph returns valid parseable state with nodes object
-  expect(graph, `[${ctx}] I1: GET /graph must return object`).toBeDefined()
-  expect(typeof graph.nodes, `[${ctx}] I1: nodes must be an object`).toBe('object')
-
-  // I2: Every API-created node still tracked must exist in graph AND on disk
-  for (const [nodeId] of tracked.nodesViaApi) {
-    if (graph.nodes[nodeId]) {
-      const exists = await fileExists(nodeId)
-      expect(exists, `[${ctx}] I2: API node ${nodeId} in graph but missing on disk`).toBe(true)
-    }
-  }
-
-  // I3: No orphan edges — every edge target must either exist in graph OR have been explicitly deleted
-  // (daemon does not cascade-delete incoming edges — this is a known behavior we document here)
-  for (const [nodeId, node] of Object.entries(graph.nodes)) {
-    if (node.outgoingEdges) {
-      for (const edge of node.outgoingEdges) {
-        const targetExists = !!graph.nodes[edge.targetId]
-        const targetWasDeleted = tracked.deletedNodeIds.has(edge.targetId)
-        expect(
-          targetExists || targetWasDeleted,
-          `[${ctx}] I3: node ${nodeId} has edge to ${edge.targetId} which never existed (not deleted, not in graph)`,
-        ).toBe(true)
-      }
-    }
-  }
-
-  // I4: Content integrity — API-created nodes should have expected content on disk
-  for (const [nodeId, trackedNode] of tracked.nodesViaApi) {
-    if (graph.nodes[nodeId]) {
-      const diskContent = await readFile(nodeId, 'utf8')
-      expect(
-        diskContent,
-        `[${ctx}] I4: node ${nodeId} disk content must contain expected body`,
-      ).toContain(trackedNode.content.trim())
-    }
-  }
-
-  // I5: Edge consistency — tracked edges should match graph state (skip deleted targets)
-  for (const [nodeId, trackedNode] of tracked.nodesViaApi) {
-    const graphNode = graph.nodes[nodeId]
-    if (graphNode && trackedNode.edges.length > 0) {
-      const graphEdgeTargets = (graphNode.outgoingEdges || []).map((e) => e.targetId)
-      for (const expectedTarget of trackedNode.edges) {
-        // Only assert edge exists if target node still exists in graph and wasn't deleted
-        if (graph.nodes[expectedTarget] && !tracked.deletedNodeIds.has(expectedTarget)) {
-          expect(
-            graphEdgeTargets,
-            `[${ctx}] I5: node ${nodeId} should have edge to ${expectedTarget}`,
-          ).toContain(expectedTarget)
-        }
-      }
-    }
-  }
-
-  // I6: Node count sanity — all tracked API nodes that exist should be in graph
-  for (const [nodeId] of tracked.nodesViaApi) {
-    const onDisk = await fileExists(nodeId)
-    if (onDisk) {
-      expect(
-        graph.nodes[nodeId],
-        `[${ctx}] I6: node ${nodeId} exists on disk but missing from graph`,
-      ).toBeDefined()
-    }
-  }
-}
-
-// ---- Fuzzer ----
+import { generateAction } from './system-lifecycle-fuzz/actions.ts'
+import { cleanupSequence } from './system-lifecycle-fuzz/drain.ts'
+import { fetchDaemonGraph } from './system-lifecycle-fuzz/daemon-http.ts'
+import { assertInvariants } from './system-lifecycle-fuzz/invariants.ts'
+import { waitFor } from './system-lifecycle-fuzz/poll.ts'
+import { mulberry32, randInt } from './system-lifecycle-fuzz/prng.ts'
+import { emptyTrackedState, resetForNextSequence } from './system-lifecycle-fuzz/types.ts'
 
 describe('system lifecycle fuzz (100 sequences, black-box HTTP)', () => {
   let root: string
@@ -434,16 +40,16 @@ describe('system lifecycle fuzz (100 sequences, black-box HTTP)', () => {
     const SEQUENCES = 100
     const topRng = mulberry32(SEED)
 
+    // tracked.deletedNodeIds accumulates across sequences so I3's
+    // `targetWasDeleted` predicate keeps holding for any leaked-then-referenced
+    // node — what used to be a cross-sequence I3 false-positive becomes a
+    // silent correct pass.
+    const tracked = emptyTrackedState()
+
     for (let seq = 0; seq < SEQUENCES; seq++) {
       const seqSeed = (topRng() * 0xFFFFFFFF) >>> 0
       const seqRng = mulberry32(seqSeed)
       const seqLen = randInt(seqRng, 8, 20)
-
-      const tracked: TrackedState = {
-        filesOnDisk: new Map(),
-        nodesViaApi: new Map(),
-        deletedNodeIds: new Set(),
-      }
 
       for (let step = 0; step < seqLen; step++) {
         const action = generateAction(seqRng, vault, baseUrl, tracked, seq, step)
@@ -451,62 +57,33 @@ describe('system lifecycle fuzz (100 sequences, black-box HTTP)', () => {
 
         await action.execute()
 
-        // For file creation, wait for the watcher to pick it up
+        // createFile is best-effort: the watcher may need more time on slow CI.
+        // Per-step invariants tolerate a missing file node.
         if (action.type === 'createFile') {
           const createdFiles = [...tracked.filesOnDisk.keys()]
           const lastFile = createdFiles[createdFiles.length - 1]
           if (lastFile) {
             await waitFor(async () => {
-              const graph = await getGraph(baseUrl)
+              const graph = await fetchDaemonGraph(baseUrl)
               return !!graph.nodes[lastFile]
-            }).catch(() => {
-              // File watcher may need more time on slow CI — invariants will catch real failures
-            })
+            }).catch(() => {})
           }
         }
 
-        // For file deletion, give watcher time to process
+        // Give the watcher a moment to drop the deleted file's node.
         if (action.type === 'deleteFile') {
           await new Promise((resolve) => setTimeout(resolve, 300))
         }
 
-        // Assert invariants after every step (not just at end of sequence)
         if (step % 3 === 0 || step === seqLen - 1) {
-          await assertInvariants(baseUrl, vault, tracked, ctx)
+          await assertInvariants(baseUrl, tracked, ctx)
         }
       }
 
-      // Final invariant check at end of sequence
       const ctx = `seq=${seq} seed=0x${seqSeed.toString(16)} final`
-      await assertInvariants(baseUrl, vault, tracked, ctx)
-
-      // Clean up between sequences: delete all API nodes and files
-      for (const nodeId of tracked.nodesViaApi.keys()) {
-        const graph = await getGraph(baseUrl)
-        if (graph.nodes[nodeId]) {
-          await fetchJson(
-            `${baseUrl}/graph/node/${encodeURIComponent(nodeId)}`,
-            { method: 'DELETE' },
-          ).catch(() => {})
-        }
-      }
-      for (const filePath of tracked.filesOnDisk.keys()) {
-        if (await fileExists(filePath)) {
-          await unlink(filePath).catch(() => {})
-        }
-      }
-
-      // Wait for cleanup to propagate then verify no state leak
-      await new Promise((resolve) => setTimeout(resolve, 200))
-      const postCleanup = await getGraph(baseUrl)
-      const remainingApiNodes = [...tracked.nodesViaApi.keys()].filter((id) => postCleanup.nodes[id])
-      if (remainingApiNodes.length > 0) {
-        // Force-clean any stragglers via endpoint
-        for (const id of remainingApiNodes) {
-          await fetchJson(`${baseUrl}/graph/node/${encodeURIComponent(id)}`, { method: 'DELETE' }).catch(() => {})
-        }
-        await new Promise((resolve) => setTimeout(resolve, 100))
-      }
+      await assertInvariants(baseUrl, tracked, ctx)
+      await cleanupSequence(baseUrl, tracked)
+      resetForNextSequence(tracked)
     }
   })
 })
